Split app from server for testing purpose
